MAINT: rename plugin to entry_points
